refactor: Notification component
