+++ src/components/organisms/NotificationsManager/NotificationsManager.tsx
@@ -1,65 +1,15 @@
 import {useCallback, useEffect} from 'react';
 
-import {Tooltip} from 'antd';
-
-import {DateTime} from 'luxon';
-
-import {AlertEnum, AlertType} from '@models/alert';
+import {AlertEnum} from '@models/alert';
 
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {seenNotifications} from '@redux/reducers/main';
 
-import NotificationMarkdown from '@molecules/NotificationMarkdown';
-
-import {useCopyToClipboard} from '@hooks/useCopyToClipboard';
-
 import {sleep} from '@utils/sleep';
 
+import Notification from './Notification';
+
 import * as S from './styled';
-
-type NotificationProps = {
-  notification: AlertType;
-  badge: JSX.Element;
-};
-
-const Notification: React.FC<NotificationProps> = props => {
-  const {notification, badge} = props;
-
-  const {createdAt, title, message, extraContentType, hasSeen, type} = notification;
-
-  const copyToClipboardMessage = `Title: ${title}. Description: ${message}.`;
-
-  const {isCopied, setCopyToClipboardState} = useCopyToClipboard(copyToClipboardMessage);
-
-  const onCopyToClipboard = () => {
-    if (isCopied) {
-      return;
-    }
-
-    setCopyToClipboardState(true);
-  };
-
-  return (
-    <S.StyledDiv key={notification.id} isNew={!hasSeen} type={type}>
-      <S.DateSpan>
-        {DateTime.fromMillis(Number(createdAt)).toRelativeCalendar()}&nbsp;
-        {DateTime.fromMillis(Number(createdAt)).toFormat('T')}
-      </S.DateSpan>
-      <S.MessageContainer>
-        <S.StatusBadge>{badge}</S.StatusBadge>
-        <Tooltip title={isCopied ? 'Copied!' : 'Copy'}>
-          <S.CopyOutlined onClick={onCopyToClipboard} />
-        </Tooltip>
-        <S.MessageBodyContainer>
-          <S.TitleSpan>{title}</S.TitleSpan>
-          <S.MessageSpan>
-            <NotificationMarkdown message={message} extraContentType={extraContentType} />
-          </S.MessageSpan>
-        </S.MessageBodyContainer>
-      </S.MessageContainer>
-    </S.StyledDiv>
-  );
-};
 
 const NotificationsManager: React.FC = () => {
   const dispatch = useAppDispatch();
